Create cmake-multi-platform.yml (#1)
Create cmake-multi-platform.yml and fixed all errors with compiling on windows

+++ src/protos/dead_pixel_map.h
@@ -13,8 +13,8 @@ struct DeadPixelMap {
     /*3*/ std::optional<float> black_level_threshold;
 
     DeadPixelMap(picoproto::Message* message) :
-        data_offset (message->GetInt64(1)),
-        data_size (message->GetInt64(2)),
+        data_offset (message->GetUInt64(1)),
+        data_size ((uint32_t)message->GetUInt64(2)),
         black_level_threshold (ToOptional(message->GetFloatArray(3)))
     {}
 };
